label_loader -> label_creator, #create_label moved
